feat: 세부수정
