+++ src/api/users.ts
@@ -1,9 +1,8 @@
-﻿import type { ApiResponse } from "@/types/api";
+import type { ApiResponse } from "@/types/api";
 import type { ClubListResponse } from "@/types/club";
 import {
   buildAuthHeaders,
   buildHeaders,
-  getApiResultType,
   getResponseMessage
 } from "@/api/common";
 
@@ -49,29 +48,6 @@ export type UpdateMyReviewPayload = {
 type UpdateMyReviewResponse = ApiResponse<{
   updatedAt: string;
 }>;
-
-export type PresignedDomain =
-  | "reviews"
-  | "facilities"
-  | "clubs"
-  | "user-profile";
-export type PresignedOperation = "PUT" | "GET";
-
-export type IssuePresignedUrlPayload = {
-  domain: PresignedDomain;
-  operation: PresignedOperation;
-  fileName: string;
-  fileType: string;
-};
-
-export type IssuePresignedUrlSuccess = {
-  url: string;
-  method: string;
-  headers?: Record<string, string>;
-  expiresIn?: number;
-};
-
-type IssuePresignedUrlResponse = ApiResponse<IssuePresignedUrlSuccess>;
 
 export type MyRegion = {
   city?: string;
@@ -133,21 +109,6 @@ export const getManagedClubs = async () => {
   }
 
   return (await res.json()) as ApiResponse<ClubListResponse>;
-};
-
-export const leaveClub = async (clubId: string) => {
-  const res = await fetch(`${API_BASE_URL}/clubs/${clubId}/join`, {
-    method: "DELETE",
-    headers: buildHeaders()
-  });
-
-  const body = await res.json().catch(() => ({}));
-
-  if (!res.ok) {
-    throw new Error(getResponseMessage(body, "Failed to leave club"));
-  }
-
-  return body as ApiResponse<Record<string, never>>;
 };
 
 // API: GET /users/me
@@ -243,51 +204,4 @@ export const updateMyReview = async (
   }
 
   return body as UpdateMyReviewResponse;
-};
-
-// API: POST /presigned-url
-export const issuePresignedUrl = async (payload: IssuePresignedUrlPayload) => {
-  const res = await fetch(`${API_BASE_URL}/presigned-url`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(payload)
-  });
-
-  const body = await res.json().catch(() => ({}));
-
-  if (!res.ok) {
-    throw new Error(getResponseMessage(body, "Failed to issue presigned url"));
-  }
-
-  // 오탈자 확인: 백엔드 응답에서 resultType 대신 resultTyle이 내려올 수 있음.
-  const resultType = getApiResultType(body);
-  if (resultType !== "SUCCESS") {
-    throw new Error(getResponseMessage(body, "Failed to issue presigned url"));
-  }
-
-  return body as IssuePresignedUrlResponse;
-};
-
-export const uploadFileToPresignedUrl = async (
-  file: File,
-  presigned: IssuePresignedUrlSuccess
-) => {
-  const uploadHeaders = {
-    ...(presigned.headers ?? {}),
-    ...(!presigned.headers?.["Content-Type"] && file.type
-      ? { "Content-Type": file.type }
-      : {})
-  };
-
-  const uploadRes = await fetch(presigned.url, {
-    method: presigned.method || "PUT",
-    headers: uploadHeaders,
-    body: file
-  });
-
-  if (!uploadRes.ok) {
-    throw new Error("Failed to upload file to storage");
-  }
-
-  return presigned.url.split("?")[0];
 };
