+++ src/components/mypage/regionOptions.ts
@@ -262,19 +262,7 @@ export const REGION_OPTIONS: RegionOption[] = [
       { id: 183, name: "\uc601\uad11\uad70" },
       { id: 184, name: "\uc7a5\uc131\uad70" },
       { id: 185, name: "\uc644\ub3c4\uad70" },
-      { id: 186, name: "\uc9c4\ub3c4" },
-      { id: 187, name: "\ubcf4\uc131\uad70" },
-      { id: 188, name: "\ud654\uc21c\uad70" },
-      { id: 189, name: "\uc7a5\ud765\uad70" },
-      { id: 190, name: "\uac15\uc9c4\uad70" },
-      { id: 191, name: "\ud574\ub0a8\uad70" },
-      { id: 192, name: "\uc601\uc554\uad70" },
-      { id: 193, name: "\ubb34\uc548\uad70" },
-      { id: 194, name: "\ud568\ud3c9\uad70" },
-      { id: 195, name: "\uc601\uad11\uad70" },
-      { id: 196, name: "\uc7a5\uc131\uad70" },
-      { id: 197, name: "\uc644\ub3c4\uad70" },
-      { id: 198, name: "\uc9c4\ub3c4\uad70" },
+      { id: 186, name: "\uc9c4\ub3c4\uad70" },
       { id: 199, name: "\uc2e0\uc548\uad70" }
     ]
   },
